+++ VendingMachine/MulledWine/DufvenkrooksWine.cs
@@ -11,7 +11,7 @@ namespace VendingMachine.MulledWine
         public override string Category { get { return "Glögg"; } }
         public override string Name { get { return "Dufvenkrooks ekologiska glögg"; } }
         public override string ProductDescription { get { return "En ekologisk glögg med 2,25 % alkohol. Volym 750 ml."; } }
-        public override decimal Price { get { return 32; } }
+        public override int Price { get { return 32; } }
 
         public void Buy()
         {
